Split Products.cs into Products.cs and ProductsOption.cs.
